+++ flink-table/flink-table-planner-blink/src/main/java/org/apache/flink/table/planner/plan/nodes/exec/processor/DeadlockBreakupProcessor.java
@@ -16,16 +16,14 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.plan.processors;
+package org.apache.flink.table.planner.plan.nodes.exec.processor;
 
 import org.apache.flink.streaming.api.transformations.ShuffleMode;
 import org.apache.flink.table.api.TableException;
 import org.apache.flink.table.planner.plan.nodes.exec.ExecEdge;
 import org.apache.flink.table.planner.plan.nodes.exec.ExecNode;
 import org.apache.flink.table.planner.plan.nodes.exec.LegacyBatchExecNode;
-import org.apache.flink.table.planner.plan.nodes.process.DAGProcessContext;
-import org.apache.flink.table.planner.plan.nodes.process.DAGProcessor;
-import org.apache.flink.table.planner.plan.processors.utils.InputPriorityConflictResolver;
+import org.apache.flink.table.planner.plan.nodes.exec.processor.utils.InputPriorityConflictResolver;
 
 import java.util.List;
